+++ Libplanet.Explorer/Queries/StateQuery.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Security.Cryptography;
 using GraphQL;
 using GraphQL.Types;
@@ -23,8 +22,8 @@ public class StateQuery
             arguments: new QueryArguments(
                 new QueryArgument<NonNullGraphType<ListGraphType<NonNullGraphType<AddressType>>>>
                     { Name = "addresses" },
-                new QueryArgument<IdGraphType> { Name = "offsetBlockHash" },
-                new QueryArgument<HashDigestSHA256Type> { Name = "offsetStateRootHash" }
+                new QueryArgument<BlockHashType> { Name = "offsetBlockHash" },
+                new QueryArgument<HashDigestType<SHA256>> { Name = "offsetStateRootHash" }
             ),
             resolve: ResolveStates
         );
@@ -33,8 +32,8 @@ public class StateQuery
             arguments: new QueryArguments(
                 new QueryArgument<NonNullGraphType<AddressType>> { Name = "owner" },
                 new QueryArgument<NonNullGraphType<CurrencyInputType>> { Name = "currency" },
-                new QueryArgument<IdGraphType> { Name = "offsetBlockHash" },
-                new QueryArgument<HashDigestSHA256Type> { Name = "offsetStateRootHash" }
+                new QueryArgument<BlockHashType> { Name = "offsetBlockHash" },
+                new QueryArgument<HashDigestType<SHA256>> { Name = "offsetStateRootHash" }
             ),
             resolve: ResolveBalance
         );
@@ -42,16 +41,16 @@ public class StateQuery
             "totalSupply",
             arguments: new QueryArguments(
                 new QueryArgument<NonNullGraphType<CurrencyInputType>> { Name = "currency" },
-                new QueryArgument<IdGraphType> { Name = "offsetBlockHash" },
-                new QueryArgument<HashDigestSHA256Type> { Name = "offsetStateRootHash" }
+                new QueryArgument<BlockHashType> { Name = "offsetBlockHash" },
+                new QueryArgument<HashDigestType<SHA256>> { Name = "offsetStateRootHash" }
             ),
             resolve: ResolveTotalSupply
         );
         Field<ListGraphType<NonNullGraphType<ValidatorType>>>(
             "validators",
             arguments: new QueryArguments(
-                new QueryArgument<IdGraphType> { Name = "offsetBlockHash" },
-                new QueryArgument<HashDigestSHA256Type> { Name = "offsetStateRootHash" }
+                new QueryArgument<BlockHashType> { Name = "offsetBlockHash" },
+                new QueryArgument<HashDigestType<SHA256>> { Name = "offsetStateRootHash" }
             ),
             resolve: ResolveValidatorSet
         );
@@ -61,7 +60,7 @@ public class StateQuery
         IResolveFieldContext<(IBlockChainStates ChainStates, IBlockPolicy Policy)> context)
     {
         Address[] addresses = context.GetArgument<Address[]>("addresses");
-        string? offsetBlockHash = context.GetArgument<string?>("offsetBlockHash");
+        BlockHash? offsetBlockHash = context.GetArgument<BlockHash?>("offsetBlockHash");
         HashDigest<SHA256>? offsetStateRootHash = context
             .GetArgument<HashDigest<SHA256>?>("offsetStateRootHash");
 
@@ -77,28 +76,15 @@ public class StateQuery
                 );
             case (blockhash: not null, _):
             {
-                BlockHash offset;
-                try
-                {
-                    offset = BlockHash.FromString(offsetBlockHash);
-                }
-                catch (Exception e)
-                {
-                    throw new ExecutionError(
-                        "offsetBlockHash must consist of hexadecimal digits.\n" + e.Message,
-                        e
-                    );
-                }
-
-                return context.Source.ChainStates.GetAccountState(
-                    offset
-                ).GetStates(addresses);
+                return context.Source.ChainStates
+                    .GetAccountState(offsetBlockHash)
+                    .GetStates(addresses);
             }
 
             case (_, srh: not null):
-                return context.Source.ChainStates.GetAccountState(
-                    offsetStateRootHash
-                ).GetStates(addresses);
+                return context.Source.ChainStates
+                    .GetAccountState(offsetStateRootHash)
+                    .GetStates(addresses);
         }
     }
 
@@ -107,7 +93,7 @@ public class StateQuery
     {
         Address owner = context.GetArgument<Address>("owner");
         Currency currency = context.GetArgument<Currency>("currency");
-        string? offsetBlockHash = context.GetArgument<string?>("offsetBlockHash");
+        BlockHash? offsetBlockHash = context.GetArgument<BlockHash?>("offsetBlockHash");
         HashDigest<SHA256>? offsetStateRootHash = context
             .GetArgument<HashDigest<SHA256>?>("offsetStateRootHash");
 
@@ -123,28 +109,15 @@ public class StateQuery
                 );
             case (blockhash: not null, _):
             {
-                BlockHash offset;
-                try
-                {
-                    offset = BlockHash.FromString(offsetBlockHash);
-                }
-                catch (Exception e)
-                {
-                    throw new ExecutionError(
-                        "offsetBlockHash must consist of hexadecimal digits.\n" + e.Message,
-                        e
-                    );
-                }
-
-                return context.Source.ChainStates.GetAccountState(
-                    offset
-                ).GetBalance(owner, currency);
+                return context.Source.ChainStates
+                    .GetAccountState(offsetBlockHash)
+                    .GetBalance(owner, currency);
             }
 
             case (_, srh: not null):
-                return context.Source.ChainStates.GetAccountState(
-                    offsetStateRootHash
-                ).GetBalance(owner, currency);
+                return context.Source.ChainStates
+                    .GetAccountState(offsetStateRootHash)
+                    .GetBalance(owner, currency);
         }
     }
 
@@ -152,7 +125,7 @@ public class StateQuery
         IResolveFieldContext<(IBlockChainStates ChainStates, IBlockPolicy Policy)> context)
     {
         Currency currency = context.GetArgument<Currency>("currency");
-        string? offsetBlockHash = context.GetArgument<string?>("offsetBlockHash");
+        BlockHash? offsetBlockHash = context.GetArgument<BlockHash?>("offsetBlockHash");
         HashDigest<SHA256>? offsetStateRootHash = context
             .GetArgument<HashDigest<SHA256>?>("offsetStateRootHash");
 
@@ -176,36 +149,20 @@ public class StateQuery
                     "Either offsetBlockHash or offsetStateRootHash must be specified."
                 );
             case (blockhash: not null, _):
-            {
-                BlockHash offset;
-                try
-                {
-                    offset = BlockHash.FromString(offsetBlockHash);
-                }
-                catch (Exception e)
-                {
-                    throw new ExecutionError(
-                        "offsetBlockHash must consist of hexadecimal digits.\n" + e.Message,
-                        e
-                    );
-                }
-
-                return context.Source.ChainStates.GetAccountState(
-                    offset
-                ).GetTotalSupply(currency);
-            }
-
+                return context.Source.ChainStates
+                    .GetAccountState(offsetBlockHash)
+                    .GetTotalSupply(currency);
             case (_, srh: not null):
-                return context.Source.ChainStates.GetAccountState(
-                    offsetStateRootHash
-                ).GetTotalSupply(currency);
+                return context.Source.ChainStates
+                    .GetAccountState(offsetStateRootHash)
+                    .GetTotalSupply(currency);
         }
     }
 
     private static object? ResolveValidatorSet(
         IResolveFieldContext<(IBlockChainStates ChainStates, IBlockPolicy Policy)> context)
     {
-        string? offsetBlockHash = context.GetArgument<string?>("offsetBlockHash");
+        BlockHash? offsetBlockHash = context.GetArgument<BlockHash?>("offsetBlockHash");
         HashDigest<SHA256>? offsetStateRootHash = context
             .GetArgument<HashDigest<SHA256>?>("offsetStateRootHash");
 
@@ -220,29 +177,13 @@ public class StateQuery
                     "Either offsetBlockHash or offsetStateRootHash must be specified."
                 );
             case (blockhash: not null, _):
-            {
-                BlockHash offset;
-                try
-                {
-                    offset = BlockHash.FromString(offsetBlockHash);
-                }
-                catch (Exception e)
-                {
-                    throw new ExecutionError(
-                        "offsetBlockHash must consist of hexadecimal digits.\n" + e.Message,
-                        e
-                    );
-                }
-
-                return context.Source.ChainStates.GetAccountState(
-                    offset
-                ).GetValidatorSet().Validators;
-            }
-
+                return context.Source.ChainStates
+                    .GetAccountState(offsetBlockHash)
+                    .GetValidatorSet().Validators;
             case (_, srh: not null):
-                return context.Source.ChainStates.GetAccountState(
-                    offsetStateRootHash
-                ).GetValidatorSet().Validators;
+                return context.Source.ChainStates
+                    .GetAccountState(offsetStateRootHash)
+                    .GetValidatorSet().Validators;
         }
     }
 }